Handling the return type of balance

+++ Project/InstaPay/src/InstaPayPackage/DatabaseHandler.java
@@ -30,7 +30,7 @@ public class DatabaseHandler {
       preparedStatement.setString(1, account.getNumberPhone());
 
       ResultSet resultSet = preparedStatement.executeQuery();
-      return resultSet.getInt("balance");
+      return resultSet.getDouble("balance");
     }
   }
